+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 
 ## Unreleased
+
+* `LaterPayClient.get_add_url()` and `LaterPayClient.get_buy_url()` accept
+  `**kwargs`.
+
 * The parameter `skip_add_to_invoice` in `LaterPayClient.get_add_url()` and 
   `LaterPayClient.get_buy_url()` is deprecated and will be removed in a future
   release.